moved fetchVideos to youtube_video_util
