+++ src/actions/search_result_actions.js
@@ -1,5 +1,3 @@
-import * as SearchResultAPI from '../util/search_result_util';
-
 import * as YoutubeVideoAPI from '../util/youtube_video_util';
 
 export const RECEIVE_VIDEOS = 'RECEIVE_VIDEOS';
@@ -15,7 +13,7 @@ export const clearVideos = () => ({
 });
 
 export const searchVideos = query => dispatch => {
-  return SearchResultAPI.fetchVideos(query).then(
+  return YoutubeVideoAPI.fetchVideos(query).then(
     res => res.json()
   ).then(
     videos => dispatch(receiveVideos(videos.items))
