+++ src/util/youtube_video_util.js
@@ -44,3 +44,13 @@ export const fetchTrending = () => {
 
   return fetch(fullUrl);
 };
+
+export const fetchVideos = query => {
+  let base_url = `https://www.googleapis.com/youtube/v3/search`;
+  let part ='snippet';
+  let type = 'video';
+
+  let full_url = `${base_url}?part=${part}&q=${query}&type=${type}&key=${YT_API_KEY.publicDataKey}`;
+
+  return fetch(full_url);
+};
